backend/requirements.txt
fastapi==0.115.2
starlette==0.38.6
uvicorn==0.30.6
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
psycopg
psycopg-binary==3.2.9
psycopg2-binary>=2.9.9  # Добавляем для совместимости
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt>=4.0.1
PyYAML>=6.0.1
